+++ src/Blazored.Video/Support/TimeRanges.cs
@@ -1,10 +1,15 @@
-﻿namespace Blazored.Video.Support
-{
-	public class TimeRanges
-	{
-		public double[] Start { get; set; }
-		public double[] End { get; set; }
-		public int Length { get; set; }
+﻿using System.Collections.Generic;
 
+namespace Blazored.Video.Support
+{
+	public class TimeRanges : List<TimeRange>
+	{
+	}
+
+	public class TimeRange
+	{
+		public int Index { get; set; }
+		public double Start { get; set; }
+		public double End { get; set; }
 	}
 }
